move to dumb and smart
