+++ acceptance/pkg/suite/steps.go
@@ -1,4 +1,4 @@
-package acceptance
+package suite
 
 import (
 	"context"
