+++ acceptance/test/dumb-proxy/config/proxy/certificate.yaml
