+++ acceptance/test/dumb-proxy/config/proxy/kustomization.yaml
