+++ acceptance/test/dumb-proxy/config/proxy/nginx.conf
