+++ acceptance/test/dumb-proxy/config/proxy/proxy.yaml
