+++ acceptance/test/dumb-proxy/config/proxy/rbac.yaml
