+++ acceptance/test/dumb-proxy/kind-config.yaml
