+++ acceptance/test/smart-proxy/.gitignore
