+++ acceptance/test/smart-proxy/config/proxy-auth/certificate.yaml
