+++ acceptance/test/smart-proxy/config/proxy-auth/kustomization.yaml
