+++ acceptance/test/smart-proxy/config/proxy-auth/nginx.conf
