+++ acceptance/test/smart-proxy/config/proxy-auth/proxy.yaml
