+++ acceptance/test/smart-proxy/config/proxy-auth/rbac.yaml
